Update the helm-operator docs for upcoming release
Please hold off on merging this until fluxcd/helm-operator#625 is merged
and Helm operator 1.4.0 is released.

The version 1.3.0 and 1.4.0 both can be released at the same time, since
we are not showing historical navigation in the legacy docs by version
anymore, I can skip the 1.3.0 bump in here.

Both versions are reflected in the changelogs, and of course both will
still get an entry in the Helm chart.

(See:
fluxcd/helm-operator#622, fluxcd/helm-operator#623
and chart releases in:
fluxcd/helm-operator#624, fluxcd/helm-operator#625
)

Signed-off-by: Kingdon Barrett <kingdon@weave.works>