Add Caution, Note and Warning callouts with i18n support
Having titles for each callout makes more work for
translation further down the line.

In some cases, it might not be appropriate to create a title, which
results in repeating the class of the warning in the title.

Moreover, when using assistive technologies like voiceover, declaring
Note, Caution, Warning at the beginning helps convey the type of note.

Signed-off-by: Alison Dowdney <alison@alisondowdney.com>