mark chart.spec.version as specific to HelmRepository source in example
In my opinion it is not visible enough that `chart.spec.version` is only considered for `HelmRepository` sources (and silently ignored otherwise.
I'm aware that the text below the example mentions this, but I think it would be benefitial for the code example itself to make that fact clear.

Adding this comment is only a suggestion (which I'm not 100% happy with myself) -- feel free to substitute a different way of pointing out the connection between `HelmRepository` and the version spec.

Signed-off-by: Tom Janson <priv.tom.janson@gmail.com>