Remove note about principle of least privilege for now
As @squaremo noted in
https://github.com/fluxcd/website/pull/641#pullrequestreview-826510951: Flux
controllers don't all need full access to all of the API, so it doesn't adhere
to the principle of least privilege, for the sake of convenience.

To-do: follow up to see if we have a list of these minimal permissions anywhere
so that users who want to follow this principle will know the minimum required.

Signed-off-by: Scott Rigby <scott@r6by.com>