DRAFT - Flux from End-to-End
This is an early draft, intended to be written in the style
of https://www.envoyproxy.io/docs/envoy/v1.19.1/intro/life_of_a_request
based on https://github.com/fluxcd/website/issues/586 – an issue in our
Flux Security Audit Follow-Up project.

I am interested in feedback about the structure, style, and goals of the
document. Welcome to reply in this PR with early feedback of any kind.

Also I am aware of the potential for overlap between this document and
those that are produced in #595, #641, and #711. I will try to keep an
eye on that effort as it develops so we are not repeating ourselves.

Signed-off-by: Kingdon Barrett <kingdon@weave.works>