Definitions
These words are common vernacular in Flux documentation. We define them
here so that their use can be understood, so they are not so unfamiliar
for any newcomers, and to ensure the meaning is unambiguously clear.

* Describe architecture of Flux using defined terms
* Add links to controller repositories
* Maybe link to the repository here, or should we link to docs instead?

Link to the docs instead of git repo

The GitHub page has the same information more-or-less but we are in the
docs, so we ought to link to docs.

* Kustomize Controller
* Helm Controller

Signed-off-by: Kingdon Barrett <kingdon@weave.works>