the squash commit
* some general reformatting

Put all of the links at the footer, and a second pass at making sure all
whitespace is reasonable and line endings are frequently placed enough.

* complete numbering cleanup
* Mark NetworkPolicy doc update as TODO

I cannot block this PR on providing a link to a section in the install
docs that we haven't written yet, and there is still more editing to do
here on this document, so I don't want to broaden this PR any further.

If the install doc is updated then we can remove the excess detail and
make the NetworkPolicy coverage in this section evergreen. For now, I
have the details written out here, in the End-to-End doc.

* Git is a proper name
* rework intro

The described goals of the document were not exactly where we landed.

* better table of contents, many changes
* fixup TODOs

Two of the todos are addressed, and links made back to open issues that will resolve them.

Signed-off-by: Kingdon Barrett <kingdon@weave.works>

all TODOs are issues now

Signed-off-by: Kingdon Barrett <kingdon@weave.works>

Reorder Image Reflector/Automation controllers

The Image Reflector comes first and Automation depends on it, so it
makes sense to describe them in chronological order like the rest of the
doc is trying to do.

Signed-off-by: Kingdon Barrett <kingdon@weave.works>

add mermaid diagrams

Signed-off-by: Kingdon Barrett <kingdon@weave.works>

add link to Get Started, elaborate purpose

The purpose of this document is elaborated a little more clearly, and
a link to the Get Started guide is placed early in the text.

Signed-off-by: Kingdon Barrett <kingdon@weave.works>

Apply suggestions from code review

Accept a batch of suggestions

Co-authored-by: Daniel Holbach <daniel@weave.works>
Co-authored-by: Paulo Gomes <paulo.gomes.uk@gmail.com>
Signed-off-by: Kingdon Barrett <kingdon@weave.works>

proof - 1/n (up to Terminology)

Signed-off-by: Kingdon Barrett <kingdon@weave.works>

proof - 2/n (up to Sources)

Signed-off-by: Kingdon Barrett <kingdon@weave.works>

say enough about editor integration for now

Signed-off-by: Kingdon Barrett <kingdon@weave.works>

proof - 3/n (up to Receivers)

Signed-off-by: Kingdon Barrett <kingdon@weave.works>

proof - 4/n (up to HelmRelease)

Signed-off-by: Kingdon Barrett <kingdon@weave.works>

proof - 5/5 (that is it!)

Signed-off-by: Kingdon Barrett <kingdon@weave.works>