GitHub Actions App Build guide
When I was reviewing some of the docs I contributed earlier and trying
to apply them to a project, I found the "Docker Build" example to be
lacking in the existing "Manifest Generation" tutorial. It does what it
needs to do for there, which is to briefly cover docker build, but only
a side-note and an example of "what we told folks to do back when Flux
v1 was widely used."

This doc shows a single GitHub Actions workflow that produces suitable
builds for BOTH git tags (ideally tags which are valid semver numbers)
AND git branches.

This example does both, so it should be a good example to highlight when
we just want to show what is a minimum CI task needed to work with Flux
that you shouldn't likely outgrow.

(squash: change title)
(squash: change introduction)
(squash: Add a reference for image provenance)
(squash: grammar)
(squash: change emphasis and grammar)
(squash: fix broken link)
(squash: implement feedback)
(squash: improved flow)
(squash: check links)

There was a missing link for [semantic version]

The other links have been reordered at the bottom to be in the order
that they appear in the text

Signed-off-by: Kingdon Barrett <kingdon@weave.works>