Make Flux UIs more obvious and show images in more than one place
This a bit of a hack, and I would appreciate some help styling this
more nicely.

Unfortunately can the gallery theme we use
(github.com/mfg92/hugo-shortcode-gallery) only show images which are in
a page bundle, so a situation like this (or similar):

directory/
    index.md
    img1.png
    img2.png

Doing this across various places would need us 1) duplicate images and
2) potentially having to update across many places as well.

I create a re-usable partial with Bootstrap's Card Group, Modal and
Carousel. Unfortunately I couldn't use Carousel controls or indicators
as the modal already uses the "data-target" argument already.

Help or feedback is welcome!

Signed-off-by: Daniel Holbach <daniel@weave.works>