GA: adapt to SC, NC and KC APIs graduating to v1
The following API types have been graduated to v1:

- NC Receiver
- KC Kustomization
- SC GitRepository

In addition to that we want the API navigation entries to appear at
the bottom of the tree. To accomplish that the import script is now
able to parse a weight parameter from imported markdown specs and put
it into the front matter of the resulting file.

refs fluxcd/notification-controller#436

Signed-off-by: Max Jonas Werner <mail@makk.es>