Mark some use-cases doc in need of refresh
The use-cases docs that are not covered here are all ones intended to be
migrated into the Installation guides, per the GSoD plan that is being
advanced here.

So the use-cases section of the docs may go away entirely, (unless some
of these docs turn out to survive all this without getting relocated):

https://github.com/fluxcd/website/issues/1560

* strengthen some language

And break out one of the longer texts into an "info" class message,
"Author's Note"

Open to suggestions about how to change this to look more Production

* ambient credentials used to secure

better language

karmada condense two disclaimers into one

I don't think we will archive the Karmada doc, since it was contributed
by an outside person and it is a CNCF Sandbox project, but I'm giving it
the same disclaimer as I'm giving all the docs as they're under review.

* Remove note from helm promotion guide that was reviewed (fixup)

Signed-off-by: Kingdon Barrett <kingdon@weave.works>

* add warning to e2e guide

Signed-off-by: Kingdon Barrett <kingdon@weave.works>