docs: Use DoesNotExist operator in shard selector
This commit updates the docs to use `DoesNotExist` operator (`!`)
for sharding selector.

The current docs use `NotIn` (`notin`) operator, which means that each
time a new shard is added, the root flux deployments needs to be updated
to exclude the new shard label. By using `DoesNotExist` operator, the
root flux deployments can remain unchanged when a new shard is added.
This makes it slightly easier to add and remove shards.

Signed-off-by: Jimmi Dyson <jimmidyson@gmail.com>