Add version dropdown and deprecation notice
I took the partial that's part of this change from the Docsy sources
and slightly modified it to have the "deprecation-warning" class and
some additional text. The original can be found here:
https://github.com/google/docsy/blob/3f3172a190ac9b24aa5b9c695f62c3fee8273ac2/layouts/partials/version-banner.html

This commit also moves the link to Flagger into the top navbar so that
we can actually use the version drop-down for listing Flux versions. A
dedicated "Documentation" link in the top navbar now directs users to
the Flux documentation.

The link to GitHub is also removed to make the navbar be closer to the
one at kubernetes.io.

Parts of this change will have to be forward-ported to `main` once 2.1
has been released. I'll create a PR for this later.

Signed-off-by: Max Jonas Werner <mail@makk.es>