Add version dropdown and deprecation notice
This is a forward-port of #1606 and #1608 in preparation of the 2.1
release.

I took the partial that's part of this change from the Docsy sources
and slightly modified it to have the "deprecation-warning" class and
some additional text. The original can be found here:
https://github.com/google/docsy/blob/3f3172a190ac9b24aa5b9c695f62c3fee8273ac2/layouts/partials/version-banner.html

This commit also moves the link to Flagger into the top navbar so that
we can actually use the version drop-down for listing Flux versions. A
dedicated "Documentation" link in the top navbar now directs users to
the Flux documentation.

The link to GitHub is also removed to make the navbar be closer to the
one at kubernetes.io.

Signed-off-by: Max Jonas Werner <mail@makk.es>