Introduce a slightly stable sort order
In contributors.py, a sort_contributors function attempts to put
contributors in order, but it does not account for the fact that a large
proportion of the contributors will have made the same number of
contributions as others (all those who made 1, 2, 3 contributions...)

Use the alphabetic sort to order those contributors in a tuple.
The alphabetic sort is reversed too, but that doesn't matter, it will
not look like people are moving haphazardly around the list anymore!

* Fix reverse alphabetic sort by name, then by contributions

The sort by contributions will override the sort by name, but then names
will still be stablely sorted (so the list will not have the appearance
of being in disorder!)

Signed-off-by: Kingdon P Barrett <kingdon@tuesdaystudios.com>
(cherry picked from commit 8183ba8388086954aaa37cf250e6fb3414b0cc49)