Accommodate missing 'organizer' field in calendar
Since yesterday the calendar at lists.cncf.io stopped including the
'organizer' field in the ICS file we use for generating the event list
on the website.

This commit makes sure that we can handle this case while generating
the website in order to fix the broken Netlify deployment.

It also simplifies the Python script a little in that it doesn't use
an intermediary dictionary that is in turn transformed to the final
dictionary that is then serialized to YAML. Now, the final dictionary
is directly generated without an intermediary one.

Signed-off-by: Max Jonas Werner <mail@makk.es>
(cherry picked from commit e648a3b68a71c3b9689ef35ee07f0a7e228c4415)