Resolve SearchIndex WARN
https://xdeb.org/post/2017/make-hugo-generate-a-json-search-index-and-json-feed/
I don't know if someone was following this blog or if it has been
updated since they followed it, but I couldn't find many references to
SearchIndex that got me closer than this one. They did only define a
searchindex for "home" not for page. I don't know if I've ever seen the
production-build without this error, (so perhaps it has never worked)

Signed-off-by: Kingdon Barrett <kingdon+github@tuesdaystudios.com>