Fix calendar timing
Do not put events at the top of the calendar that started more than 1
hour ago

The website only refreshes the calendar every 6 hours. So if the event
started more than 1 hour ago, it's stale and should not be listed at the
top as a "next event" or "upcoming"

I used 50 minutes ago as "hour_ago" - there is a bit of not sure about
why this happens, whether it's our code or the recurring ical events
module that includes these events. It doesn't really matter if it's two
minutes ago or an hour ago, but I wanted to avoid "the event is starting
now and it just dropped off the calendar, how are people meant to join"

I'm sure that was the intent of including events that started earlier
today, but it really causes quite a lot of confusion from people who
only read as far as "Next event:" and found a date and time many hours
in the past.

Signed-off-by: Kingdon Barrett <kingdon+github@tuesdaystudios.com>
(cherry picked from commit c4419be1e5c58846f32708f4a70322cdf66e9d80)