flatten layouts/shortcodes/home/calendar.html
In order to use the shortcode in the external-imports, we need to avoid
having parts of it converted into <pre> sections. Hugo does this I guess
(it's an annoying change, but this is the only external-import with any
shortcode that I'm aware of, and shortcodes aren't very big)

Other changes will also be needed like some text changes and formatting.
The CSS rules to put the calendar into dark mode need to be revisited.

Signed-off-by: Kingdon Barrett <kingdon+github@tuesdaystudios.com>
(cherry picked from commit 130926ad8b57e477c44f96926bbfa486786ca288)